Небольшой рефакторинг отрисовки событий
